requirements.txt
amqp==2.3.2
beautifulsoup4==4.6.3
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
darkskylib==0.3.91
Django==1.11
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extra-views==0.12.0
django-livereload-server==0.3.2
django-polymorphic==2.0.3
django-polymorphic-auth==0.4
django-shapeshifter==18.9.23
future==0.17.1
idna==2.8
kombu>=4.2.2
pbr==5.1.1
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.11.0
sqlparse==0.2.4
stevedore==1.30.0
tornado==5.1.1
urllib3==1.24.1
vine==1.1.4
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
gunicorn
django-heroku